third_party/blink/web_tests/fast/dom/shadow/drop-event-in-shadow.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/event-sender-util.js"></script>
<script src="../../../resources/js-test.js"></script>
<script>
function createBox(name) {
    var div = document.createElement('div');
    div.style.width = '100px';
    div.style.height = '100px';

    div.addEventListener('dragover', function(e) {
        e.preventDefault();
    });
    div.addEventListener('drop', function(e) {
        debug('PASS: drop event is fired.');
        e.preventDefault();
    });

    return div;
}

function runTest() {
    var shadowRoot = host.attachShadow({mode: 'open'});
    var shadowDiv = createBox('shadow');
    shadowRoot.appendChild(shadowDiv);

    eventSender.beginDragWithFiles(["../resources/apple.gif"]);
    mouseMoveToElem(shadowDiv);
    eventSender.mouseUp();
}
</script>
</head>
<body onload="runTest()">

<div id="host"></div>
<pre id="console"></pre>
</body>
</html>
