third_party/blink/web_tests/fast/dom/shadow/drop-event-in-shadow-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS: drop event is fired.
